Fix non-deterministic copy_image_revert_lifecycle test
This test is failing fairly often in zuul. The test attempts to delete
one of the stores while the import is happening to simulate a failure
and ensure that the first store is not reverted. This can race and fail
to delete before the task completes. Instead, if we just delete the
store directory before the import, we know that when the flow gets to
the second store's step, it will fail to write the image, giving the same
result, reliably.

Change-Id: Ibeea8a93f140f35daa47024d07ad806162ff690d