Add functional test for task status updating
This was really hard to test before the introduction of the
SynchronousAPIBase test class, so this is added after that patch for
convenience.

This tests uses two mocks, but it's otherwise quite "functional" in its
realness. It mocks out the timer so it always fires, and it mocks out
the effective call to glance_store and uses that as a hook to grab the
task state in lockstep (and avoid making another copy of the fake data
on disk, since we need to generate a couple MiB of data to test this).

Change-Id: Ibd0c802efa2723e11ece1097c9bf1ad68b1a820c