Move SynchronousAPIBase to a generalized location
The base class for the tests added in test_images_import_locking provides
a mechanism to make API cals directly against the WSGI stack, without
starting a separate server and using the local networking. This is useful
for cases where fault injection of global state needs to be altered, where
this is very difficult in the existing fork-and-exec functional test
model.

This moves that test base class out to the functional module, expands the
documentation a little, and also generalizes the request methods for
wider applicability.

Change-Id: I59e3b5d5d4b69f076092b9950c0d34467a6636ad
(cherry picked from commit 36cbc50e7d70699d6ca5dd4fa81844567e39ddf0)