Make test_copy_image_revert_lifecycle handle 409 on import retry
This test is very racy in general, and specifically on the post-revert
retry to do the import, it may start the next import before the previous
one has finished reverting. This patch makes it retry a few times with
delay if an HTTP 409 is received so it is tolerant of that situation once
it becomes possible in the subsequent patch to add import locking.

Change-Id: Ic933f170d43b290fd819e8527ecb60be7f7f3f89
(cherry picked from commit 8c7342cbc15fdbcbafa1ead9aa183e45862cf177)