Adjust jobs for devstack WSGI mode default
Devstack is switching the default glance operational model back to pure-WSGI
mode for parity with the rest of the services. This adjusts our jobs to fit
with that so that we're testing the default, but also the standalone case.
After this change, we end up with the following for the vanilla devstack
jobs:

 - The "-import" job, which takes the devstack default of WSGI mode,
   with all the tweaks required for import to be enabled. It configures
   import conversion to raw, metadata injection, and confirms that with a
   post-run check for that metadata.
 - The "-import-standalone" job, which inherits all the behaviors from
   the job above, but configures glance in standalone mode.

Depends-On: https://review.opendev.org/#/c/742884
Change-Id: Ia3fb8aba83bbd7a1399aef136ce8857b14d08435