Make web-download revert all stores on fail
If web-download fails to stage the image, no subsequent import to
store tasks could have run. Thus, we should remove all of them from
"importing" and add them to "failed" so that it is obvious from the
outside that we failed.

This is another good reason for having the $image/tasks API, but we
also should not continue advertising "yes we're importing to $store"
when we are not.

NOTE(lyarwood): This includes a partial squashed revert of the
.zuul.yaml changes in Ifd55c44fef4e2187052d77084dc9c0fa9c9a0d16 that
removed nova-ceph-multistore. This is reintroduced here to ensure this
change resolves the stable/victoria specific bug #1919993 and the
original bug #1914826.

Change-Id: Iebbb2dcb767ecf3c965f34f1ca04af20a2039be1
Closes-Bug: #1914826
Closes-Bug: #1919993
(cherry picked from commit cbb2af6e329c1bf6c1c39cb59f95aaad2c0d7a07)