Add missing fail case tests for image_conversion
The tests for the image_conversion plugin were missing basically all
of the places where the task can fail (or NOP). This adds tests for
those things to validate behavior before moving the task to use the
import wrapper.

Note that in the process, I found a potential bug in the handling of
the metadata response from qemu-img when probing for the format. We
tolerate "format" being missing from the result, but will end up
passing None as an argument to 'qemu-img convert', which will fail.
This fixes that to raise RuntimeError in line with the existing
behavior of "raise RuntimeError if any error is reported".

Change-Id: I8f1d03275e6ec51a802cc4b4107f3ab648f535a1