Fix nonsensical test mocks and assertions
The test_inject_image_metadata test was doing some completely
nonsensical mock behavior, like mocking something after we have
run the method under test, and re-mocking the image repo with the
thing that was supposed to change so that the assertions were not
even checking the actual behavior. None of the three tests were
actually testing the desired behavior.

In the case of the admin_user test, it was also asserting that things
were not run that *should* be, probably because the mocks were done
after the execution and asserting that they *were* run would never
work.

This also wires up the image and image_repo mocks that were in
setUp(), but unused (and incorrect) so that we can assert the calls
are actually as we expect.

Change-Id: Ifc2f49d49ff62c49c7dd1d2174a78bb1451b7bf3