Add a test for migration naming and phase rules
Any time we have a migration for a given release, we want to make sure
that we have a definition for each phase, even if it is empty. This
adds a test that makes sure we conform to that in the future, to help
reduce reviewer load and requiring the tribal knowledge at the point
of review.

While doing this, this also validates that the migration files fit
a general pattern of release_phaseNN_description.py format.

Change-Id: Iaea8a81b64503ad82525eb8edd609253ff1d47b9