trivial: remove unnecessary grouping in base policies
We've broken basic policies into granular checks with simple names and
we use them to construct more complex checks. In that process we
accidentally added some additional nesting to two of the check strings,
which isn't necessary.

This commit updates the check strings to remove an extra set of
parenthesis.

Change-Id: Iafa37d64a9779a3b646c34f328c62dfd6cd3e7f3