Fix erroneous exit from copy wait loop
The wait_for_copying() helper will exit the loop if the *last*
store is found in the list, instead of *all* of them. This technically
works if the stores are processed in the same order we are checking,
but it's fragile and likely to fail in confusing ways.

This makes us only exit if all of them are present.

Change-Id: I8d9ba50f46e22b6740fdbdec6f8ef7c61dddbcf1