Ignore stale image property removal
Occasionally we fail to delete a property from an image and get a
StaleDataError from SQLAlchemy. It looks like maybe we are just racing
with another operation to delete a single property (i.e. a race
between image delete and the end of an import cleanup).

This ignores that "can't delete because it is already deleted" error
and logs to debug about the situation for later auditing and
confirmation.

Change-Id: Ida5bc659b7481fa0a7f3926a54520628a7a4c406
Closes-Bug: #1895173