Cinder Store: Update legacy images tests
update tests for cinder's new attachment API code in glance
cinder store. This leaves a couple of mocks in place for the old
way of doing things so that we can tolerate the old and proposed
changes to glance_store. We can remove those once we require the
newer glance_store version.

Change-Id: If4db20affbba37806f8419b6b30050a32bcc2435