Add user_get_staging_usage() to DB API
This adds a method that counts the sizes of images that are in the
process of doing an import, indicating that they are using staging
storage resources. We gather images that are in obvious states of
being created via import, as well as any images that appear to be
active but in the middle of a copy, since those could be using staging
resources as well.

This is somewhat mirrored from user_get_storage_usage(). Since there
was no test for that method, this adds a test case for both that and
the new method.

Partially-implements: blueprint glance-unified-quotas

Change-Id: Ic2b045549ac0d6112b0c565d2e993576a5626799