Refactor gateway get_repo auth layer
This makes us able to optionally not get the authorization layer on
top of the stack. The authorization layer makes everything on an
Image read-only based on hard-coded checks for "owner or admin".
If we are going to move to using policy to enforce modifications
to images, we need to discard this layer.

When the flag is True we also don't insert the policy layer in the
middle of the stack.

This patch cleans up the stacking to be a little more modular, and
adds a flag to control whether or not we get the authorization layer
on top or not.

Partially-Implements: blueprint policy-refactor

Change-Id: Ie8bea199175a0eb35428b1508af1c49111a87c37