Reproduce bug #1932337
This reproduces the behavior described in bug 1932337, where a post-
upgrade GET by a non-owning user will return an error for a valid image
because the cinder store migration cannot complete.

This asserts the broken behavior to reproduce the problem, and the
rest of the test will be enabled when the bug is fixed.

Note that the test for this was duplicating code from the base class,
and since some changes were needed to those methods, this removes the
duplication and modifies only the base.

Change-Id: I741e6377b3aa40c14d637f0fbdf396dc468f6ebd