Move lazy store update to locations layer
The authorization layer was hijacked (IMHO) to provide the location
metadata lazy-loading. Since we are disabling that layer in the
policy refactor, we need somewhere else to do that. The best place
to do it seems to be in the actual locations layer of the onion,
so this patch moves it there.

No tests are changed in this patch because this is already confirmed
from the images resource, this just moves where it's done, and not
needing any test change shows that this doesn't change any externally-
visible behavior.

Change-Id: I70372c44f4b2d9547cb8c40b43df17d1085ef03c