Fix broken test_update_queued_image_with_hidden
This fixes the test to behave the way we expect. It was failing to
do the update because it was using an image the requester did not
own, and asserting the found behavior of 403. However, the intent
here is to allow it to be updated. So, this uses the proper image and
asserts the proper behavior.

Change-Id: I71afe6a877485c8f92e67dcf32bb475c1a1a42a3
Closes-Bug: #1933360