Fix failing copy_image flow init
The new networkx requirement apparently causes us to fail the
test_init_copy_flow_as_non_owner test because a None value can not be
added to the flow graph. This just mocks out the glance_store call
that is failing to get the staging store to avoid us choking there
so we can proceed. We are just testing get_flow so we never use it
anyway.

Change-Id: I6fcb14ad240105ed0b1b9086c6c6c95034b4dd42