Add check_is_image_mutable() legacy helper
This adds a method to let API code replicate the legacy admin-or-owner
behavior where needed. Since the authorization layer does this in
some places (which is being removed) and the DB does it in others,
we need a way to replicate that check when secure_rbac is not in use
for certain code paths (like image delete) which need it.

Change-Id: I89785ae04e7f66651a1b9673f7e5513c2eac10fc
Partially-implements: blueprint policy-refactor