Resolve compatibility with oslo.db future
Change I468048c01455d9dfe12004736c181806dc218e37 in oslo.db means that
the table returned from 'oslo_db.utils.get_table' is no longer bound to
a 'Connection' or 'Engine'. Get out ahead of this breaking change by
creating out own 'Connection' as part of the test. Note that we use
'Connection.execute' instead of 'Engine.execute' as the latter is
deprecated in SQLAlchemy 1.4 for removal in 2.0.

Related-Bug: #1939716

Change-Id: I957d2a9d7c875120bed0022ce1e953b2ec9f97cd
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>