Check download_image policy in the API
Note that this adds a clause to the download exception handler which
returns 404 if the user can't download the image via policy, but leaves
the existing internal forbidden->403 handler because of how image
property protections work.

Also Cache middleware do not have any influence of policy refactoring.

Partially-Implements: blueprint policy-refactor

Change-Id: I060f22145d2915bfd0b0b3efc2912c24e89b9a30