Remove duplicate namespace utilities in functional tests
We're in the process of adding test coverage for the metadef API, which
is sorely needed. This commit introduces two different base classes that
help consolidate some duplication that's grown across the metadef tests.

Since the metadef API contains a whole bunch of nest resources, the
first class is a set of utilities that are handy for creating
metadef resources. This is nice because almost every resources in the
metadef API needs a namespace.

The second class is a subclass of the functional test case, but provides
a layer for metadef-specific test utilities, like assertions. We can
continue using this layer to consolidate test utilities without bloating
the base functional test class.

Change-Id: Ic469ee50b9a3f162fd75a8290766d1e0bb4b26f5