Use single quotes in image policies
The oslopolicy-sample-generator wraps strings in double quotes, which
breaks with the glance sample policy file since it's using double quotes
to evaluate strings.

This commit fixes the issue so that we're using single quotes in the
check strings so oslo.policy can wrap them with double-quotes and it
won't break yaml parsers.

Change-Id: I4f4553d91e58047c05ea853a8eb3764376338a45